//maddy boot

+++ src/main/java/com/ats/rusasoft/model/FacultyConference.java
@@ -24,7 +24,7 @@ public class FacultyConference {
 	private Date confDate;
 	private String confVenue;
 	private String 	confFundFrom;
-	private int confFundAmt;
+	private float confFundAmt;
 	private int delStatus;
 	private int isActive;
 	private int makerUserId;
@@ -80,10 +80,10 @@ public class FacultyConference {
 	public void setConfFundFrom(String confFundFrom) {
 		this.confFundFrom = confFundFrom;
 	}
-	public int getConfFundAmt() {
+	public float getConfFundAmt() {
 		return confFundAmt;
 	}
-	public void setConfFundAmt(int confFundAmt) {
+	public void setConfFundAmt(float confFundAmt) {
 		this.confFundAmt = confFundAmt;
 	}
 	public int getDelStatus() {
